[cell_fabric] Common primitive code goes into cell_fabric
