+++ align/cell_fabric/__init__.py
@@ -2,7 +2,6 @@ from .grid import *
 from .generators import *
 from .canvas import Canvas
 from .default import DefaultCanvas
-from .primitive import DefaultPrimitiveGenerator
 from .pdk import Pdk
 
 from .primitive import generate_primitive
