+++ align/cell_fabric/primitive.py
@@ -5,326 +5,14 @@ import logging
 import collections
 import math
 import json
+import importlib.util
 
 from .generators import Wire
 from . import gen_lef
 from . import gen_gds_json
+from .pdk import Pdk
 
 logger = logging.getLogger(__name__)
-
-class DefaultPrimitiveGenerator():
-
-    def _addMOS( self, x, y, name='M1', reflect=False, **parameters):
-
-        fullname = f'{name}_X{x}_Y{y}'
-        self.subinsts[fullname].parameters.update(parameters)
-
-        def _connect_diffusion(i, pin):
-            self.addWire( self.m1, None, None, i, (grid_y0, -1), (grid_y1, 1))
-            for j in range(((self.finDummy+3)//2), self.v0.h_clg.n):
-                self.addVia( self.v0, f'{fullname}:{pin}', None, i, (y, j))
-            self._xpins[name][pin].append(i)
-
-        # Draw FEOL Layers
-
-        self.addWire( self.active, None, None, y, (x,1), (x+1,-1))
-        self.addWire( self.pc, None, None, y, (x,1), (x+1,-1))
-        for i in range(1,  self.finsPerUnitCell):
-            self.addWire( self.fin, None, None,  self.finsPerUnitCell*y+i, x, x+1)
-        for i in range(self.gatesPerUnitCell):
-            self.addWire( self.pl, None, None, self.gatesPerUnitCell*x+i,   (y,0), (y,1))
-
-        # Source, Drain, Gate Connections
-
-        grid_y0 = y*self.m2PerUnitCell + self.finDummy//2-1
-        grid_y1 = grid_y0+(self.finsPerUnitCell - 2*self.finDummy + 2)//2
-        gate_x = x * self.gatesPerUnitCell + self.gatesPerUnitCell // 2
-        # Connect Gate (gate_x)
-        self.addWire( self.m1, None, None, gate_x , (grid_y0, -1), (grid_y1, 1))
-        self.addVia( self.va, f'{fullname}:G', None, gate_x, (y*self.m2PerUnitCell//2, 1))
-        self._xpins[name]['G'].append(gate_x)
-        # Connect Source & Drain
-        if reflect:
-            _connect_diffusion(gate_x + 1, 'S') #S
-            _connect_diffusion(gate_x - 1, 'D') #D
-        else:
-            _connect_diffusion(gate_x - 1, 'S') #S
-            _connect_diffusion(gate_x + 1, 'D') #D
-
-    def _connectDevicePins(self, y, connections):
-        center_track = y * self.m2PerUnitCell + self.m2PerUnitCell // 2 # Used for m1 extension
-        for track, (net, conn) in enumerate(connections.items(), start=1):
-            contacts = {track for inst, pins in self._xpins.items()
-                              for pin, m1tracks in pins.items()
-                              for track in m1tracks if (inst, pin) in conn}
-            for j in range(self.minvias):
-                current_track = y * self.m2PerUnitCell + len(connections) * j + track
-                self.addWireAndViaSet(net, None, self.m2, self.v1, current_track, contacts)
-                self._nets[net][current_track] = contacts
-                # Extend m1 if needed. TODO: Should we draw longer M1s to begin with?
-                direction = 1 if current_track > center_track else -1
-                for i in contacts:
-                    self.addWire( self.m1, net, None, i, (center_track, -1 * direction), (current_track, direction))
-
-
-    def _connectNets(self, x_cells, y_cells): 
-
-        def _get_wire_terminators(intersecting_tracks):
-            minx, maxx = min(intersecting_tracks), max(intersecting_tracks)
-            # BEGIN: Quick & dirty MinL DRC error fix.
-            minL = 2 # TODO: Make this MinL dependent
-            L = maxx - minx
-            if center_track - minL // 2 <= minx and maxx <= center_track + minL // 2:
-                minx, maxx = (center_track - minL // 2, center_track + minL // 2)
-            elif L < minL:
-                minx, maxx = (minx - (minL - L), maxx) if minx >= center_track else (minx, maxx + (minL - L))
-            # END: Quick & dirty MinL DRC error fix.
-            return (minx, maxx)
-
-        center_track = (x_cells * self.gatesPerUnitCell) // 2
-        m3start = (x_cells * self.gatesPerUnitCell - len(self._nets) * self.minvias) // 2
-        for track, (net, conn) in enumerate(self._nets.items(), start=1):
-            for j in range(self.minvias):
-                current_track = m3start + len(self._nets) * j + track
-                contacts = conn.keys()
-                if len(contacts) == 1: # Create m2 terminal
-                    i = next(iter(contacts))
-                    minx, maxx = _get_wire_terminators(conn[i])
-                    self.addWire(self.m2, net, net, i, (minx, -1), (maxx, 1))
-                else: # create m3 terminal(s)
-                    self.addWireAndViaSet(net, net, self.m3, self.v2, current_track, contacts)
-                    # Extend m2 if needed. TODO: What to do if we go beyond cell boundary?
-                    for i, locs in conn.items():
-                        minx, maxx = _get_wire_terminators([*locs, current_track])
-                        self.addWire(self.m2, net, None, i, (minx, -1), (maxx, 1))
-
-        self.addWire( self.m2, 'B', 'B', (y_cells)* self.m2PerUnitCell + self.lFin//4, (0, 1), (x_cells*self.gatesPerUnitCell, -1))
-
-    def _addBodyContact(self, x, y, yloc=None, name='M1'):
-        fullname = f'{name}_X{x}_Y{y}'
-        if yloc is not None:
-            y = yloc
-        h = self.m2PerUnitCell
-        gu = self.gatesPerUnitCell
-        gate_x = x*gu + gu // 2
-        self._xpins[name]['B'].append(gate_x)
-        self.addWire( self.activeb, None, None, y, (x,1), (x+1,-1))
-        self.addWire( self.pb, None, None, y, (x,1), (x+1,-1))
-        self.addWire( self.m1, None, None, gate_x, ((y+1)*h+3, -1), ((y+1)*h+self.lFin//2-3, 1))
-        self.addWire( self.LISDb, None, None, gate_x, ((y+1)*h+3, -1), ((y+1)*h+self.lFin//2-3, 1))
-        self.addVia( self.va, f'{fullname}:B', None, gate_x, (y+1)*h + self.lFin//4)
-        self.addVia( self.v1, 'B', None, gate_x, (y+1)*h + self.lFin//4)
-        for i in range(self.finsPerUnitCell, self.finsPerUnitCell+self.lFin):
-            self.addWire( self.fin, None, None,  self.finsPerUnitCell*y+i, x, x+1)
-
-    def _addMOSArray( self, x_cells, y_cells, pattern, connections, minvias = 1, **parameters):
-        if minvias * len(connections) > self.m2PerUnitCell - 1:
-            self.minvias = (self.m2PerUnitCell - 1) // len(connections)
-            logger.warning( f"Using minvias = {self.minvias}. Cannot route {len(connections)} signals using minvias = {minvias} (max m2 / unit cell = {self.m2PerUnitCell})" )
-        else:
-            self.minvias = minvias
-        names = ['M1'] if pattern == 0 else ['M1', 'M2']
-        self._nets = collections.defaultdict(lambda: collections.defaultdict(list)) # net:m2track:m1contacts (Updated by self._connectDevicePins)
-        for y in range(y_cells):
-            self._xpins = collections.defaultdict(lambda: collections.defaultdict(list)) # inst:pin:m1tracks (Updated by self._addMOS)
-            for x in range(x_cells):
-                if pattern == 0: # None (single transistor)
-                    # TODO: Not sure this works without dummies. Currently:
-                    # A A A A A A
-                    self._addMOS(x, y, names[0], False, **parameters)
-                    self._addBodyContact(x, y, y_cells - 1, names[0])
-                elif pattern == 1: # CC
-                    # TODO: Think this can be improved. Currently:
-                    # A B B A A' B' B' A'
-                    # B A A B B' A' A' B'
-                    # A B B A A' B' B' A'
-                    self._addMOS(x, y, names[((x // 2) % 2 + x % 2 + (y % 2)) % 2], x >= x_cells // 2, **parameters)
-                    self._addBodyContact(x, y, y_cells - 1, names[((x // 2) % 2 + x % 2 + (y % 2)) % 2])
-                elif pattern == 2: # interdigitated
-                    # TODO: Evaluate if this is truly interdigitated. Currently:
-                    # A B A B A B
-                    # B A B A B A
-                    # A B A B A B
-                    self._addMOS(x, y, names[((x % 2) + (y % 2)) % 2], False, **parameters)
-                    self._addBodyContact(x, y, y_cells - 1, names[((x % 2) + (y % 2)) % 2])
-                elif pattern == 3: # CurrentMirror
-                    # TODO: Evaluate if this needs to change. Currently:
-                    # B B B A A B B B
-                    # B B B A A B B B
-                    self._addMOS(x, y, names[0 if 0 <= ((x_cells // 2) - x) <= 1 else 1], False, **parameters)
-                    self._addBodyContact(x, y, y_cells - 1, names[0 if 0 <= ((x_cells // 2) - x) <= 1 else 1])
-                else:
-                    assert False, "Unknown pattern"
-            self._connectDevicePins(y, connections)
-        self._connectNets(x_cells, y_cells) 
-
-    def addNMOSArray( self, x_cells, y_cells, pattern, connections, **parameters):
-
-        self._addMOSArray(x_cells, y_cells, pattern, connections, **parameters)
-
-        #####   Nselect Placement   #####
-        self.addRegion( self.nselect, None, None, (0, -1), 0, (x_cells*self.gatesPerUnitCell, -1), y_cells* self.finsPerUnitCell+self.lFin) 
-
-    def addPMOSArray( self, x_cells, y_cells, pattern, connections, **parameters):
-
-        self._addMOSArray(x_cells, y_cells, pattern, connections, **parameters)
-
-        #####   Pselect and Nwell Placement   #####
-        self.addRegion( self.pselect, None, None, (0, -1), 0, (x_cells*self.gatesPerUnitCell, -1), y_cells* self.finsPerUnitCell+self.lFin)
-        self.addRegion( self.nwell, None, None, (0, -1), 0, (x_cells*self.gatesPerUnitCell, -1), y_cells* self.finsPerUnitCell+self.lFin)
-
-    def addCap( self, unit_cap):
-
-        x_length = float((math.sqrt(unit_cap/2))*1000)
-        y_length = float((math.sqrt(unit_cap/2))*1000)  
-
-        c_m1_p = self.pdk['Cap']['m1Pitch']
-        c_m1_w = self.pdk['Cap']['m1Width']
-        c_m2_p = self.pdk['Cap']['m2Pitch']
-        c_m2_w = self.pdk['Cap']['m2Width']
-        m1_p = self.pdk['M1']['Pitch']
-        m2_p = self.pdk['M2']['Pitch']
-
-        logger.debug( f"Pitches {c_m1_p} {c_m2_p} {m1_p} {m2_p}")
-
-        def compute( l, p, w):
-            # this is nonsense but if l is a multiple of 2p say 2kp, then 2kp+p-w/(2p) is always k
-            return int( 2*round(  (l+p-w)/(2.0*p) ))
-
-        x_number = compute( x_length, c_m1_p, c_m1_w)
-        y_number = compute( y_length, c_m2_p, c_m2_w)
-
-        logger.debug( f"Number of wires {x_number} {y_number}")
-
-        def roundup( x, p):
-            return (x+p-1)//p
-
-        last_y1_track = roundup( (y_number-1)*c_m2_p, m2_p)
-        last_x1_track = roundup( (x_number-1)*c_m1_p, m1_p)
-
-        grid_y0 = 0
-        grid_y1 = grid_y0 + last_y1_track
-
-        for i in range(x_number-1):
-            grid_x = i
-            net = 'PLUS' if i%2 == 1 else 'MINUS'
-            self.addWire( self.m1n, net, None, grid_x, (grid_y0, -1), (grid_y1, 1))
-            self.addWire( self.m3n, net, None, grid_x, (grid_y0, -1), (grid_y1, 1))
-
-            grid_y = ((i+1)%2)*grid_y1
-
-            self.addVia( self.v1_nx, net, None, grid_x, grid_y)
-            self.addVia( self.v2_nx, net, None, grid_x, grid_y)
-
-        pin = 'PLUS'
-        # Don't port m1 per Yaguang instructions
-        self.addWire( self.m1, 'PLUS', None, last_x1_track, (grid_y0, -1), (grid_y1, 1))
-        # don't port m3 (or port one or the other)
-        self.addWire( self.m3, 'PLUS', None, last_x1_track, (grid_y0, -1), (grid_y1, 1))
-
-        grid_x0 = 0
-        grid_x1 = grid_x0 + last_x1_track
-
-        for i in range(y_number-1):
-            grid_x = ((i+1)%2)*grid_x1
-            net = 'PLUS' if i%2 == 0 else 'MINUS'
-            self.addVia( self.v1_xn, net, None, grid_x, i)
-            self.addVia( self.v2_xn, net, None, grid_x, i)
-            pin = 'PLUS' if i == 0 else None
-            self.addWire( self.m2n, net, pin, i, (grid_x0, -1), (grid_x1, 1))
-
-        pin = 'MINUS'
-        self.addWire( self.m2, 'MINUS', pin, last_y1_track, (grid_x0, -1), (grid_x1, 1))
-
-        self.addRegion( self.boundary, 'Boundary', None,
-                        0, 0,
-                        last_x1_track,
-                        last_y1_track)
-
-        #self.addRegion( self.Cboundary, 'Cboundary', None,
-        #                    -1, -1,
-        #                    last_x_track  + x * grid_cell_x_pitch + 1 + p,
-        #                    last_y1_track + y * grid_cell_y_pitch + 1)
-
-        logger.debug( f"Computed Boundary: {self.terminals[-1]} {self.terminals[-1]['rect'][2]} {self.terminals[-1]['rect'][2]%80}")
-
-    def addResArray(self, x_cells, y_cells, height, unit_res):
-
-        for x in range(x_cells):
-            for y in range(y_cells):
-                self._addRes(x, y, height, unit_res, (x == x_cells-1) and (y == y_cells-1))
-
-    def _addRes( self, x, y, height, unit_res, draw_boundary=True):
-
-        y_length = self.finsPerUnitCell * self.pdk['Fin']['Pitch'] * height
-        assert y_length != 0, (self.finsPerUnitCell, self.pdk['Fin']['Pitch'], height)
-        res_per_length = 67
-        x_number = int(round(((1000*unit_res)/(res_per_length*y_length))))
-        assert x_number >= 1, (unit_res, res_per_length, y_length)
-
-        # ga = 2 if x_number == 1 else 1 ## when number of wires is 2 then large spacing req. so contact can be placed without a DRC error 
-        # x_length = (x_number - 1) *ga*self.pdk['Cap']['m1Pitch']
-
-        y_number = int(2 *round(((y_length+self.pdk['Cap']['m2Pitch']-self.pdk['Cap']['m2Width'])/(2.0*self.pdk['Cap']['m2Pitch']))))
-
-        last_y1_track = ((y_number-1)*self.pdk['Cap']['m2Pitch']+self.pdk['M2']['Pitch']-1)//self.pdk['M2']['Pitch']
-        last_x_track = x_number - 1
-
-        m2factor = 2 ### number of m2-tracks (m2factor-1)in between two unitcells in y-direction
-        m1factor = 3
-
-        if (y_number-1) % 2 != last_y1_track % 2:
-            last_y1_track += 1 # so the last color is compatible with the external view of the cell
-
-        if last_y1_track % 2 == 1:
-            m2factor += 1 # so colors match in arrayed blocks
-
-        grid_cell_x_pitch = m1factor + last_x_track
-        grid_cell_y_pitch = m2factor + last_y1_track
-
-        grid_y0 = y*grid_cell_y_pitch
-        grid_y1 = grid_y0 + last_y1_track
-
-        for i in range(x_number):
-            (k, p) = (2*i, 1) if x_number==2 else (i, 0)
-            grid_x = k + x*grid_cell_x_pitch
-
-            self.addWire( self.m1res, None, None, grid_x, (grid_y0, -1), (grid_y1, 1))
-            if i < x_number-1:
-                grid_yh = ((i+1)%2)*last_y1_track
-                self.addWire( self.m1res2, None, None, grid_yh, (i, -1), (i+p+1, 1))
-
-#
-# Build the narrow m2 pitch grid starting at grid_cell_y_pitch*y in standard m2 pitch grids (m2.clg)
-#
-        m2n = Wire( self.m2res2.nm, self.m2res2.layer, self.m2res2.direction,
-                    clg=self.m2res2.clg.copyShift( self.m2res.clg.value( grid_cell_y_pitch*y)[0]),
-                    spg=self.m2res2.spg)
-
-        #v1n = Via( 'v1', 'via1', h_clg=m2n.clg, v_clg=self.m1res.clg)
-        #v2n = Via( 'v2', 'via2', h_clg=m2n.clg, v_clg=self.m3res.clg)
-
-        grid_x0 = x*grid_cell_x_pitch
-        grid_x1 = grid_x0 + last_x_track
-        grid_y = (x_number%2)*last_y1_track
-
-        pin = 'PLUS'
-        self.addWire( m2n, 'PLUS', pin, 0, (0, -1), (0, 1))
-        self.addVia( self.v1res, None, None, 0, 0)
-        pin = 'MINUS'
-        self.addWire( self.m2res, 'MINUS', pin, grid_y, (grid_x1+p, -1), (grid_x1+p, 1))
-        self.addVia( self.v1res, None, None, grid_x1+p, grid_y)
-
-        if draw_boundary:
-            self.addRegion( self.boundary, 'boundary', None,
-                            -1, -1,
-                            last_x_track  + x * grid_cell_x_pitch + 1 + p,
-                            last_y1_track + y * grid_cell_y_pitch + 1)
-
-            self.addRegion( self.Rboundary, 'Rboundary', None,
-                            -1, -1,
-                            last_x_track  + x * grid_cell_x_pitch + 1 + p,
-                            last_y1_track + y * grid_cell_y_pitch + 1)
 
 def get_xcells_pattern( primitive, pattern, x_cells):
     if any(primitive.startswith(f'{x}_') for x in ["CM", "CMFB"]):
@@ -347,8 +35,18 @@ def get_parameters(primitive, parameters, nfin):
     return parameters
 
 # TODO: Pass cell_pin and pattern to this function to begin with
-def generate_MOS_primitive(uc, block_name, primitive, height, nfin, x_cells, y_cells, pattern, parameters, pinswitch):
+def generate_MOS_primitive(pdkdir, block_name, primitive, height, nfin, x_cells, y_cells, pattern, parameters, pinswitch):
 
+    pdk = Pdk().load(pdkdir / 'layers.json')
+    generator = get_generator('MOSGenerator', pdkdir)
+
+    # TODO: THIS SHOULD NOT BE NEEDED !!!
+    fin = height
+    gateDummy = 3 ### Total Dummy gates per unit cell: 2*gateDummy
+    finDummy = 4  ### Total Dummy fins per unit cell: 2*finDummy
+    gate = 2
+
+    uc = generator(pdk, fin, finDummy, gate, gateDummy)
     x_cells, pattern = get_xcells_pattern(primitive, pattern, x_cells)
     parameters = get_parameters(primitive, parameters, nfin)
 
@@ -434,39 +132,49 @@ def generate_MOS_primitive(uc, block_name, primitive, height, nfin, x_cells, y_c
 
     return uc, cell_pin
 
-def generate_Cap(uc, block_name, unit_cap):
+def generate_Cap(pdkdir, block_name, unit_cap):
+
+    pdk = Pdk().load(pdkdir / 'layers.json')
+    generator = get_generator('CapGenerator', pdkdir)
+
+    uc = generator(pdk)
+
     uc.addCap(unit_cap)
 
     return uc, ['PLUS', 'MINUS']
 
-def generate_Res(uc, block_name, x_cells, y_cells, height, unit_res):
-    uc.addResArray(x_cells, y_cells, height, unit_res)
+def generate_Res(pdkdir, block_name, height, x_cells, y_cells, nfin, unit_res):
+
+    pdk = Pdk().load(pdkdir / 'layers.json')
+    generator = get_generator('ResGenerator', pdkdir)
+
+    fin = height
+    finDummy = 4  ### Total Dummy fins per unit cell: 2*finDummy
+
+    uc = generator(pdk, fin, finDummy)
+
+    uc.addResArray(x_cells, y_cells, nfin, unit_res)
 
     return uc, ['PLUS', 'MINUS']
+
+def get_generator(name, pdkdir):
+    spec = importlib.util.spec_from_file_location("primitive", pdkdir / 'primitive.py')
+    primitive = importlib.util.module_from_spec(spec)
+    spec.loader.exec_module(primitive)
+    return getattr(primitive, name)
 
 # WARNING: Bad code. Changing these default values breaks functionality.
 def generate_primitive(block_name, primitive, height=12, x_cells=1, y_cells=1, pattern=1, value=12, parameters=None, pinswitch=0, pdkdir=pathlib.Path.cwd(), outputdir=pathlib.Path.cwd()):
 
     assert pdkdir.exists() and pdkdir.is_dir(), "PDK directory does not exist"
-    sys.path.insert(0, str(pdkdir))
-    from primitive import PrimitiveGenerator
-    sys.path.pop(0)
-
-    # TODO: THIS SHOULD NOT BE NEEDED !!!
-    fin = height
-    gateDummy = 3 ### Total Dummy gates per unit cell: 2*gateDummy
-    finDummy = 4  ### Total Dummy fins per unit cell: 2*finDummy
-    gate = 2
-
-    uc = PrimitiveGenerator( fin, finDummy, gate, gateDummy)
 
     if 'MOS' in primitive:
-        uc, cell_pin = generate_MOS_primitive(uc, block_name, primitive, height, value, x_cells, y_cells, pattern, parameters, pinswitch)
+        uc, cell_pin = generate_MOS_primitive(pdkdir, block_name, primitive, height, value, x_cells, y_cells, pattern, parameters, pinswitch)
     elif 'cap' in primitive.lower():
-        uc, cell_pin = generate_Cap(uc, block_name, value)
+        uc, cell_pin = generate_Cap(pdkdir, block_name, value)
         uc.setBboxFromBoundary()
     elif 'Res' in primitive:
-        uc, cell_pin = generate_Res(uc, block_name, x_cells, y_cells, value[0], value[1])
+        uc, cell_pin = generate_Res(pdkdir, block_name, height, x_cells, y_cells, value[0], value[1])
         uc.setBboxFromBoundary()
     else:
         raise NotImplementedError(f"Unrecognized primitive {primitive}")
