+++ pdks/FinFET14nm_Mock_PDK/primitive.py
@@ -1,10 +1,9 @@
-from canvas import FinFET14nm_Mock_PDK_Canvas
-from align.cell_fabric import DefaultPrimitiveGenerator
+from align.cell_fabric import default
 
-import logging
-logger = logging.getLogger(__name__)
+# Override default MOSGenerator._addMOS & _addBodyContact 
+# (to add fin & LISD layers)
 
-class PrimitiveGenerator(FinFET14nm_Mock_PDK_Canvas, DefaultPrimitiveGenerator):
+class MOSGenerator(default.MOSGenerator):
 
     def _addMOS( self, x, y, name='M1', reflect=False, **parameters):
 
@@ -12,8 +11,27 @@ class PrimitiveGenerator(FinFET14nm_Mock_PDK_Canvas, DefaultPrimitiveGenerator):
         super()._addMOS(x, y, name, reflect, **parameters)
 
         # Draw Technology Specific Layers
-        self.addWire( self.RVT,  None, None, y,          (x, 1), (x+1, -1))
+        for i in range(1,  self.finsPerUnitCell):
+            self.addWire( self.fin, None, None,  self.finsPerUnitCell*y+i, x, x+1)
 
         gate_x = x * self.gatesPerUnitCell + self.gatesPerUnitCell // 2
         self.addWire( self.LISD, None, None, gate_x - 1, (y, 1), (y+1, -1))
         self.addWire( self.LISD, None, None, gate_x + 1, (y, 1), (y+1, -1))
+
+    def _addBodyContact(self, x, y, yloc=None, name='M1'):
+        super()._addBodyContact(x, y, yloc, name)
+        if yloc is not None:
+            y = yloc
+        h = self.m2PerUnitCell
+        gu = self.gatesPerUnitCell
+        gate_x = x*gu + gu // 2
+        self.addWire( self.LISDb, None, None, gate_x, ((y+1)*h+3, -1), ((y+1)*h+self.lFin//2-3, 1))
+        for i in range(self.finsPerUnitCell, self.finsPerUnitCell+self.lFin):
+            self.addWire( self.fin, None, None,  self.finsPerUnitCell*y+i, x, x+1)
+
+#
+# Default Cap & Res generators are good enough
+#
+
+CapGenerator = default.CapGenerator
+ResGenerator = default.ResGenerator
